restructuring native type operators
