+++ include/universal/native/native
@@ -1,0 +1,16 @@
+#pragma once
+// native standard header
+//
+// Copyright (C) 2017-2020 Stillwater Supercomputing, Inc.
+//
+// This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#ifndef _NATIVE_
+#define _NATIVE_
+
+#include <universal/native/bit_functions.hpp>
+#include <universal/native/byteArray.hpp>
+#include <universal/native/integers.hpp>
+#include <universal/native/ieee-754.hpp>
+#include <universal/native/manipulators.hpp>
+
+#endif
