+++ tests/numerical/thin_triangle.cpp
@@ -1,6 +1,6 @@
 ﻿// goldberg_thin_triangle.cpp: example program showing the Goldberg thin triangle example
 //
-// Copyright (C) 2017-2019 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2020 Stillwater Supercomputing, Inc.
 //
 // This file is part of the UNIVERSAL project, which is released under an MIT Open Source license.
 #include "common.hpp"
